feat(projects/products/unstable/repo_contract_enforcer/backend): update rules and tests
