+++ projects/products/unstable/repo_contract_enforcer/backend/src/rules/crate_rules.rs
@@ -1,3 +1,6 @@
+//! projects/products/unstable/repo_contract_enforcer/backend/src/rules/crate_rules.rs
+use std::{collections, fs, path};
+
 use crate::scan::file_scanner::FileScanner;
 use crate::scan::rust_parser::{MainItemViolationKind, RustParser};
 use crate::{config, reports, rules};
@@ -16,7 +19,7 @@ const EXCLUDED_PRIMARY_STEMS: [&str; 4] = ["main", "mod", "lib", "public_api"];
 
 impl CrateRules {
     pub fn evaluate(
-        product_dir: &std::path::Path,
+        product_dir: &path::Path,
         product_name: &str,
         scope: config::path_classification::PathClassification,
         mode: config::enforcement_mode::EnforcementMode,
@@ -84,14 +87,24 @@ impl CrateRules {
                     ));
                 }
                 if txt.contains("[[bin]]") {
-                    out.push(make_violation(
-                        RuleId::Crate,
-                        ViolationCode::CrateNotBinOnly,
-                        (scope, mode),
-                        &cargo,
-                        "crate must be bin-only (no [[bin]])",
-                        (true, None),
-                    ));
+                    let cli_name = read_cli_name_from_metadata(crate_path);
+                    let bin_name = parse_bin_name_from_cargo(&txt);
+
+                    let allowed = match (&cli_name, &bin_name) {
+                        (Some(cli), Some(bin)) => cli == bin,
+                        _ => false,
+                    };
+
+                    if !allowed {
+                        out.push(make_violation(
+                            RuleId::Crate,
+                            ViolationCode::CrateNotBinOnly,
+                            (scope, mode),
+                            &cargo,
+                            "crate must be bin-only (no [[bin]]); to declare a CLI name, set cli_name in metadata.ron and use [[bin]] name = cli_name",
+                            (true, None),
+                        ));
+                    }
                 }
                 if cargo_declares_dependency(&txt, "serde_json") {
                     out.push(make_violation(
@@ -305,8 +318,8 @@ impl CrateRules {
 }
 
 fn should_enforce_primary_item_contract(
-    src_dir: &std::path::Path,
-    rs_file: &std::path::Path,
+    src_dir: &path::Path,
+    rs_file: &path::Path,
     stem: &str,
 ) -> bool {
     if is_excluded_primary_stem(stem) {
@@ -321,7 +334,7 @@ fn should_enforce_primary_item_contract(
     !comps.any(|c| c.as_os_str() == "tests")
 }
 
-fn is_file_under_tests_dir(src_dir: &std::path::Path, rs_file: &std::path::Path) -> bool {
+fn is_file_under_tests_dir(src_dir: &path::Path, rs_file: &path::Path) -> bool {
     let rel = match rs_file.strip_prefix(src_dir) {
         Ok(path) => path,
         Err(_) => return false,
@@ -329,8 +342,8 @@ fn is_file_under_tests_dir(src_dir: &std::path::Path, rs_file: &std::path::Path)
     rel.components().any(|c| c.as_os_str() == "tests")
 }
 
-fn allowed_inline_test_attr_lines_for_mod_rs(content: &str) -> std::collections::HashSet<u32> {
-    let mut allowed = std::collections::HashSet::new();
+fn allowed_inline_test_attr_lines_for_mod_rs(content: &str) -> collections::HashSet<u32> {
+    let mut allowed = collections::HashSet::new();
     let lines: Vec<&str> = content.lines().collect();
     for (idx, line) in lines.iter().enumerate() {
         if line.trim() != "#[cfg(test)]" {
@@ -351,7 +364,7 @@ fn allowed_inline_test_attr_lines_for_mod_rs(content: &str) -> std::collections:
     allowed
 }
 
-fn expected_paired_test_path(rs_file: &std::path::Path) -> Option<std::path::PathBuf> {
+fn expected_paired_test_path(rs_file: &path::Path) -> Option<path::PathBuf> {
     let stem = rs_file
         .file_stem()
         .and_then(|s| s.to_str())
@@ -400,7 +413,7 @@ fn make_violation(
     rule_id: rules::rule_id::RuleId,
     code: reports::violation_code::ViolationCode,
     context: RuleContext,
-    path: &std::path::Path,
+    path: &path::Path,
     message: &str,
     meta: ViolationMeta,
 ) -> reports::violation::Violation {
@@ -432,7 +445,7 @@ fn make_violation(
 fn read_text_or_emit_violation(
     out: &mut Vec<reports::violation::Violation>,
     context: RuleContext,
-    path: &std::path::Path,
+    path: &path::Path,
     kind: &str,
 ) -> Option<String> {
     match std::fs::read_to_string(path) {
@@ -494,537 +507,76 @@ fn cargo_declares_dependency(cargo_toml_content: &str, crate_name: &str) -> bool
     false
 }
 
-#[cfg(test)]
-mod tests {
-    use super::CrateRules;
-    use crate::config::enforcement_mode::EnforcementMode;
-    use crate::config::path_classification::PathClassification;
-    use crate::reports::violation_code::ViolationCode;
-    use std::fs;
-    use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
+/// Reads `cli_name` from `metadata.ron` located at the crate root.
+/// Returns `None` if the file is absent, unreadable, or `cli_name` is not set.
+fn read_cli_name_from_metadata(crate_path: &path::Path) -> Option<String> {
+    let metadata_path = crate_path.join("metadata.ron");
+    let content = fs::read_to_string(&metadata_path).ok()?;
 
-    fn temp_product_root() -> PathBuf {
-        let stamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("time before epoch")
-            .as_nanos();
-        let root = std::env::temp_dir().join(format!("repo_contract_enforcer_crate_rules_{stamp}"));
-        fs::create_dir_all(&root).expect("create temp product root");
-        root
+    for line in content.lines() {
+        let trimmed = line.trim();
+
+        // Match `cli_name:` key anywhere on the line.
+        let rest = trimmed.strip_prefix("cli_name:").or_else(|| {
+            // Handle inline tuple: `..., cli_name: Some("va"), ...`
+            let pos = trimmed.find("cli_name:")?;
+            Some(&trimmed[pos + "cli_name:".len()..])
+        });
+
+        if let Some(rest) = rest {
+            // Strip trailing comma, whitespace, comments.
+            let value = rest
+                .trim()
+                .trim_end_matches(',')
+                .trim()
+                .split("//") // strip inline comments
+                .next()
+                .unwrap_or_default()
+                .trim();
+
+            // Match `Some("...")` — tolerant of inner whitespace.
+            let inner = value
+                .strip_prefix("Some(")
+                .or_else(|| value.strip_prefix("Some ("))?
+                .trim()
+                .strip_prefix('"')?;
+
+            let name = inner.split('"').next()?.trim();
+
+            if !name.is_empty() {
+                return Some(name.to_string());
+            }
+
+            // Explicit None — stop here.
+            return None;
+        }
     }
+    None
+}
 
-    fn write_minimal_bin_crate(crate_root: &std::path::Path, crate_name: &str) {
-        fs::create_dir_all(crate_root.join("src")).expect("create crate src dir");
-        fs::write(
-            crate_root.join("Cargo.toml"),
-            format!(
-                "[package]\nname = \"{crate_name}\"\nversion = \"0.1.0\"\nedition = \"2021\"\n"
-            ),
-        )
-        .expect("write Cargo.toml");
-        fs::write(crate_root.join("src/main.rs"), "fn main() {}\n").expect("write main.rs");
+/// Extracts the `name` field from the first `[[bin]]` section in a Cargo.toml string.
+/// Returns `None` if `[[bin]]` is absent or has no `name` field.
+fn parse_bin_name_from_cargo(cargo_toml: &str) -> Option<String> {
+    let mut in_bin_section = false;
+    for line in cargo_toml.lines() {
+        let trimmed = line.trim();
+        if trimmed == "[[bin]]" {
+            in_bin_section = true;
+            continue;
+        }
+        // New section starts — exit bin section
+        if in_bin_section && trimmed.starts_with('[') {
+            break;
+        }
+        if in_bin_section && let Some(rest) = trimmed.strip_prefix("name") {
+            let rest = rest.trim();
+            if let Some(value) = rest.strip_prefix('=') {
+                let raw = value.trim().trim_matches('"');
+                if !raw.is_empty() {
+                    return Some(raw.to_string());
+                }
+            }
+        }
     }
-
-    #[test]
-    fn crate_requires_paired_test_file_for_logic_module() {
-        let product_root = temp_product_root();
-        let product_name = "project_alpha";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_alpha_backend");
-        write_minimal_bin_crate(&ui, "project_alpha_ui");
-
-        let request_rs = backend.join("src/protocol/request.rs");
-        fs::create_dir_all(request_rs.parent().expect("request parent"))
-            .expect("mkdir request parent");
-        fs::write(
-            &request_rs,
-            "pub struct Request;\n\nimpl Request {\n    pub fn is_valid(&self) -> bool { true }\n}\n",
-        )
-        .expect("write request.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateMissingPairedTestFile
-                && v.path.ends_with("backend/src/protocol/request.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_paired_test_file_contract_accepts_matching_tests_path() {
-        let product_root = temp_product_root();
-        let product_name = "project_beta";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_beta_backend");
-        write_minimal_bin_crate(&ui, "project_beta_ui");
-
-        let request_rs = backend.join("src/protocol/request.rs");
-        fs::create_dir_all(request_rs.parent().expect("request parent"))
-            .expect("mkdir request parent");
-        fs::write(
-            &request_rs,
-            "pub struct Request;\n\nimpl Request {\n    pub fn is_valid(&self) -> bool { true }\n}\n",
-        )
-        .expect("write request.rs");
-
-        let request_test = backend.join("src/protocol/tests/request.rs");
-        fs::create_dir_all(request_test.parent().expect("request test parent"))
-            .expect("mkdir request test parent");
-        fs::write(&request_test, "#[test]\nfn request_contract() {}\n")
-            .expect("write request test");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(!violations.iter().any(|v| {
-            matches!(
-                v.violation_code,
-                ViolationCode::CrateMissingPairedTestFile
-                    | ViolationCode::CratePairedTestNotUnitStyle
-            )
-        }));
-    }
-
-    #[test]
-    fn crate_paired_test_file_requires_unit_test_marker() {
-        let product_root = temp_product_root();
-        let product_name = "project_gamma";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_gamma_backend");
-        write_minimal_bin_crate(&ui, "project_gamma_ui");
-
-        let request_rs = backend.join("src/protocol/request.rs");
-        fs::create_dir_all(request_rs.parent().expect("request parent"))
-            .expect("mkdir request parent");
-        fs::write(
-            &request_rs,
-            "pub struct Request;\n\nimpl Request {\n    pub fn is_valid(&self) -> bool { true }\n}\n",
-        )
-        .expect("write request.rs");
-
-        let request_test = backend.join("src/protocol/tests/request.rs");
-        fs::create_dir_all(request_test.parent().expect("request test parent"))
-            .expect("mkdir request test parent");
-        fs::write(
-            &request_test,
-            "pub fn helper_for_external_integration() -> bool { true }\n",
-        )
-        .expect("write request test");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CratePairedTestNotUnitStyle
-                && v.path.ends_with("backend/src/protocol/tests/request.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_data_only_module_does_not_require_paired_test_file() {
-        let product_root = temp_product_root();
-        let product_name = "project_delta";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_delta_backend");
-        write_minimal_bin_crate(&ui, "project_delta_ui");
-
-        let record_rs = backend.join("src/store/account_record.rs");
-        fs::create_dir_all(record_rs.parent().expect("record parent"))
-            .expect("mkdir record parent");
-        fs::write(
-            &record_rs,
-            "#[derive(Debug, Clone)]\npub struct AccountRecord {\n    pub id: String,\n}\n",
-        )
-        .expect("write account_record.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(!violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateMissingPairedTestFile
-                && v.path.ends_with("backend/src/store/account_record.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_detects_unscoped_pub_in_binary_main_module() {
-        let product_root = temp_product_root();
-        let product_name = "project_eps";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_eps_backend");
-        write_minimal_bin_crate(&ui, "project_eps_ui");
-
-        fs::write(
-            backend.join("src/main.rs"),
-            "#![allow(dead_code)]\npub mod public_api;\nfn main() {}\n",
-        )
-        .expect("write backend main.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateBinaryMainUnscopedPub
-                && v.path.ends_with("backend/src/main.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_detects_disallowed_items_in_binary_main_module() {
-        let product_root = temp_product_root();
-        let product_name = "project_zeta";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_zeta_backend");
-        write_minimal_bin_crate(&ui, "project_zeta_ui");
-
-        fs::write(
-            backend.join("src/main.rs"),
-            "pub(crate) struct App;\nenum Mode { Fast }\ntrait Runner {}\nimpl App { fn run(&self) {} }\nfn helper() {}\nfn main() {}\n",
-        )
-        .expect("write backend main.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(
-            violations
-                .iter()
-                .any(|v| v.violation_code == ViolationCode::CrateBinaryMainContainsStruct)
-        );
-        assert!(
-            violations
-                .iter()
-                .any(|v| v.violation_code == ViolationCode::CrateBinaryMainContainsEnum)
-        );
-        assert!(
-            violations
-                .iter()
-                .any(|v| v.violation_code == ViolationCode::CrateBinaryMainContainsTrait)
-        );
-        assert!(
-            violations
-                .iter()
-                .any(|v| v.violation_code == ViolationCode::CrateBinaryMainContainsImpl)
-        );
-        assert!(
-            violations
-                .iter()
-                .any(|v| v.violation_code == ViolationCode::CrateBinaryMainContainsNonEntrypointFn)
-        );
-    }
-
-    #[test]
-    fn crate_detects_local_use_statement_in_function_scope() {
-        let product_root = temp_product_root();
-        let product_name = "project_eta";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_eta_backend");
-        write_minimal_bin_crate(&ui, "project_eta_ui");
-
-        fs::write(
-            backend.join("src/service.rs"),
-            "pub fn run() {\n    use std::collections::HashMap;\n    let _m: HashMap<String, String> = HashMap::new();\n}\n",
-        )
-        .expect("write service.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateContainsLocalUseStatement
-                && v.path.ends_with("backend/src/service.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_detects_inline_test_attribute_outside_tests_dir() {
-        let product_root = temp_product_root();
-        let product_name = "project_theta";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_theta_backend");
-        write_minimal_bin_crate(&ui, "project_theta_ui");
-
-        fs::write(
-            backend.join("src/service.rs"),
-            "#[cfg(test)]\nmod tests {\n    #[test]\n    fn smoke() {}\n}\n",
-        )
-        .expect("write service.rs");
-        let service_test = backend.join("src/tests/service.rs");
-        fs::create_dir_all(service_test.parent().expect("service test parent"))
-            .expect("create service test parent");
-        fs::write(&service_test, "#[test]\nfn allowed_in_tests_folder() {}\n")
-            .expect("write tests/service.rs");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateContainsInlineTestAttribute
-                && v.path.ends_with("backend/src/service.rs")
-        }));
-        assert!(!violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateContainsInlineTestAttribute
-                && v.path.ends_with("backend/src/tests/service.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_allows_cfg_test_mod_tests_declaration_in_mod_rs() {
-        let product_root = temp_product_root();
-        let product_name = "project_theta_bis";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_theta_bis_backend");
-        write_minimal_bin_crate(&ui, "project_theta_bis_ui");
-
-        let domain_mod = backend.join("src/domain/mod.rs");
-        fs::create_dir_all(domain_mod.parent().expect("domain parent"))
-            .expect("create domain parent");
-        fs::write(&domain_mod, "#[cfg(test)]\nmod tests;\n").expect("write domain/mod.rs");
-
-        let domain_tests_mod = backend.join("src/domain/tests/mod.rs");
-        fs::create_dir_all(domain_tests_mod.parent().expect("domain tests parent"))
-            .expect("create domain tests parent");
-        fs::write(&domain_tests_mod, "#[test]\nfn smoke() {}\n").expect("write domain tests");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(!violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateContainsInlineTestAttribute
-                && v.path.ends_with("backend/src/domain/mod.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_allows_cfg_test_mod_tests_declaration_in_main_rs() {
-        let product_root = temp_product_root();
-        let product_name = "project_theta_ter";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_theta_ter_backend");
-        write_minimal_bin_crate(&ui, "project_theta_ter_ui");
-
-        fs::write(
-            backend.join("src/main.rs"),
-            "#[cfg(test)]\nmod tests;\nfn main() {}\n",
-        )
-        .expect("write backend main.rs");
-
-        let backend_tests_mod = backend.join("src/tests/mod.rs");
-        fs::create_dir_all(backend_tests_mod.parent().expect("backend tests parent"))
-            .expect("create backend tests parent");
-        fs::write(
-            &backend_tests_mod,
-            "#![cfg(test)]\n#[test]\nfn smoke() {}\n",
-        )
-        .expect("write backend tests mod");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(!violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateContainsInlineTestAttribute
-                && v.path.ends_with("backend/src/main.rs")
-        }));
-    }
-
-    #[test]
-    fn crate_rejects_explicit_additional_bin_targets() {
-        let product_root = temp_product_root();
-        let product_name = "project_iota";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_iota_backend");
-        write_minimal_bin_crate(&ui, "project_iota_ui");
-
-        fs::write(
-            backend.join("Cargo.toml"),
-            "[package]\nname = \"project_iota_backend\"\nversion = \"0.1.0\"\nedition = \"2021\"\n\n[[bin]]\nname = \"project_iota_backend\"\npath = \"src/main.rs\"\n\n[[bin]]\nname = \"extra\"\npath = \"src/bin/extra.rs\"\n",
-        )
-        .expect("write backend Cargo with [[bin]]");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateNotBinOnly
-                && v.path.ends_with("backend/Cargo.toml")
-                && v.message.contains("no [[bin]]")
-        }));
-    }
-
-    #[test]
-    fn crate_rejects_src_bin_targets() {
-        let product_root = temp_product_root();
-        let product_name = "project_kappa";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_kappa_backend");
-        write_minimal_bin_crate(&ui, "project_kappa_ui");
-
-        let extra = backend.join("src/bin/extra.rs");
-        fs::create_dir_all(extra.parent().expect("extra parent")).expect("mkdir extra bin parent");
-        fs::write(&extra, "fn main() {}\n").expect("write extra bin");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateNotBinOnly
-                && v.path.ends_with("backend/src/bin/extra.rs")
-                && v.message.contains("src/bin targets are forbidden")
-        }));
-    }
-
-    #[test]
-    fn crate_rejects_serde_json_dependency_for_products() {
-        let product_root = temp_product_root();
-        let product_name = "project_lambda";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_lambda_backend");
-        write_minimal_bin_crate(&ui, "project_lambda_ui");
-
-        fs::write(
-            backend.join("Cargo.toml"),
-            "[package]\nname = \"project_lambda_backend\"\nversion = \"0.1.0\"\nedition = \"2021\"\n\n[dependencies]\nserde_json = \"1\"\n",
-        )
-        .expect("write backend Cargo with serde_json");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateForbiddenSerdeJsonDependency
-                && v.path.ends_with("backend/Cargo.toml")
-        }));
-    }
-
-    #[test]
-    fn crate_requires_dioxus_dependency_in_ui_crate() {
-        let product_root = temp_product_root();
-        let product_name = "project_mu";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_mu_backend");
-        write_minimal_bin_crate(&ui, "project_mu_ui");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateUiMissingDioxusDependency
-                && v.path.ends_with("ui/Cargo.toml")
-        }));
-    }
-
-    #[test]
-    fn crate_accepts_dioxus_dependency_in_target_specific_section() {
-        let product_root = temp_product_root();
-        let product_name = "project_nu";
-
-        let backend = product_root.join("backend");
-        let ui = product_root.join("ui");
-        write_minimal_bin_crate(&backend, "project_nu_backend");
-        write_minimal_bin_crate(&ui, "project_nu_ui");
-
-        fs::write(
-            ui.join("Cargo.toml"),
-            "[package]\nname = \"project_nu_ui\"\nversion = \"0.1.0\"\nedition = \"2021\"\n\n[target.'cfg(target_arch = \"wasm32\")'.dependencies]\ndioxus = { version = \"0.7.3\", features = [\"web\"] }\n",
-        )
-        .expect("write ui Cargo with target dioxus");
-
-        let violations = CrateRules::evaluate(
-            &product_root,
-            product_name,
-            PathClassification::Stable,
-            EnforcementMode::Strict,
-        );
-
-        assert!(!violations.iter().any(|v| {
-            v.violation_code == ViolationCode::CrateUiMissingDioxusDependency
-                && v.path.ends_with("ui/Cargo.toml")
-        }));
-    }
+    None
 }
