+++ projects/products/unstable/repo_contract_enforcer/backend/src/rules/mod.rs
@@ -7,3 +7,6 @@ pub mod rule_engine;
 pub mod rule_id;
 pub mod structure_rules;
 pub mod tool_rules;
+
+#[cfg(test)]
+mod tests;
