+++ projects/products/unstable/repo_contract_enforcer/backend/src/rules/tests/mod.rs
@@ -1,0 +1,2 @@
+mod crate_rules;
+mod tool_rules;
